requirements.txt
Pillow==7.2.0
pyTelegramBotAPI==3.7.3
requests~=2.24.0